+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-04-03-1
+
+Changes:
+
+- Request abort signal
+
 ## v2023-03-20-1
 
 Changes: